fix loading rubric list from content service
Content service was unmarshalling rubric list from object model found in item-renderer project and could not be loaded by
the teacher handscoring application.
