+++ pom.xml
@@ -32,7 +32,6 @@
         <tds-common.version>5.0.3</tds-common.version>
         <tds-common-legacy.version>5.0.3</tds-common-legacy.version>
         <tds-shared-build.version>4.0.6.RELEASE</tds-shared-build.version>
-        <item-renderer.version>4.0.7.RELEASE</item-renderer.version>
         <jackson.version>2.9.2</jackson.version>
         <spring.saml.version>1.0.4.RELEASE</spring.saml.version>
 
@@ -59,26 +58,7 @@
                 <artifactId>shared-spring</artifactId>
                 <version>${tds-shared-build.version}</version>
             </dependency>
-            <dependency>
-                <groupId>org.opentestsystem.delivery</groupId>
-                <artifactId>item-renderer</artifactId>
-                <version>${item-renderer.version}</version>
-                <exclusions>
-                    <exclusion>
-                        <groupId>org.springframework</groupId>
-                        <artifactId>spring-context</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>org.springframework</groupId>
-                        <artifactId>spring-mock</artifactId>
-                    </exclusion>
-                    <exclusion>
-                        <groupId>org.apache.tomcat</groupId>
-                        <artifactId>el-api</artifactId>
-                    </exclusion>
-                </exclusions>
-            </dependency>
-
+            
             <!-- Spring -->
             <dependency>
                 <groupId>org.springframework</groupId>
@@ -96,6 +76,12 @@
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-starter-security</artifactId>
                 <version>1.5.8.RELEASE</version>
+            </dependency>
+            <dependency>
+                <groupId>org.springframework.security</groupId>
+                <artifactId>spring-security-test</artifactId>
+                <version>5.0.5.RELEASE</version>                
+                <scope>test</scope>
             </dependency>
             <dependency>
                 <groupId>org.springframework.security.extensions</groupId>
