+++ service/pom.xml
@@ -40,10 +40,6 @@
             <groupId>org.opentestsystem.shared</groupId>
             <artifactId>shared-spring</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.opentestsystem.delivery</groupId>
-            <artifactId>item-renderer</artifactId>
-        </dependency>
 
         <!-- Spring -->
         <dependency>
@@ -94,7 +90,10 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
-            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.security</groupId>
+            <artifactId>spring-security-test</artifactId>
         </dependency>
         <dependency>
             <groupId>org.assertj</groupId>
